bites_rc_v7/requirements.txt
celery>=5.3.0
redis>=5.0.0
django-redis>=5.3.0
supabase
requests
Pillow
numpy>=1.26.4,<2
openai-whisper
torch
git+https://github.com/openai/CLIP.git
yt-dlp>=2023.3.4 